requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
neo4j==5.18.0
requests==2.31.0
pandas==2.2.2
rich==13.7.1
python-dotenv==1.0.1
